evaluate/pyproject.toml
[project]
name = "hlab-evaluate"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["torch==2.3.*","torchvision==0.18.*","numpy>=1.26","pandas>=2.2","pyyaml>=6.0","scikit-learn>=1.4"]
